Hacky fix to keep bullet's frame counter consistent.
A bullet's frame counter needs to be updated _after_ running its Lua
function to keep the mental model of the bullet consistent. For example,
when we initially run a bullet, we expect the turn counter to start at 0
since technically, no frames have elapsed for that bullet.

As such, since the frame counter determines bullet state, when we kill a
bullet, we want the frame counter to be one behind the state we want it
to be in. I.e. after the function is run, the turn counter is
immediately updated and will be flagged as dead (or dying) at the
correct time.